Split router sub-menus to a separate file
