+++ themes-default/slim/src/router/routes.js
@@ -1,79 +1,9 @@
-const showSubMenu = function() {
-    const { $route, $store } = this;
-    const { config, notifiers } = $store.state;
-
-    const indexerName = $route.params.indexer || $route.query.indexername;
-    const showId = $route.params.id || $route.query.seriesid;
-
-    const show = $store.getters.getCurrentShow;
-    const { showQueueStatus } = show;
-
-    const queuedActionStatus = action => {
-        if (!showQueueStatus) {
-            return false;
-        }
-        return Boolean(showQueueStatus.find(status => status.action === action && status.active === true));
-    };
-
-    const isBeingAdded = queuedActionStatus('isBeingAdded');
-    const isBeingUpdated = queuedActionStatus('isBeingUpdated');
-    const isBeingSubtitled = queuedActionStatus('isBeingSubtitled');
-
-    let menu = [{
-        title: 'Edit',
-        path: `home/editShow?indexername=${indexerName}&seriesid=${showId}`,
-        icon: 'ui-icon ui-icon-pencil'
-    }];
-    if (!isBeingAdded && !isBeingUpdated) {
-        menu = menu.concat([
-            {
-                title: show.config.paused ? 'Resume' : 'Pause',
-                path: `home/togglePause?indexername=${indexerName}&seriesid=${showId}`,
-                icon: `ui-icon ui-icon-${show.config.paused ? 'play' : 'pause'}`
-            },
-            {
-                title: 'Remove',
-                path: `home/deleteShow?indexername=${indexerName}&seriesid=${showId}`,
-                confirm: 'removeshow',
-                icon: 'ui-icon ui-icon-trash'
-            },
-            {
-                title: 'Re-scan files',
-                path: `home/refreshShow?indexername=${indexerName}&seriesid=${showId}`,
-                icon: 'ui-icon ui-icon-refresh'
-            },
-            {
-                title: 'Force Full Update',
-                path: `home/updateShow?indexername=${indexerName}&seriesid=${showId}`,
-                icon: 'ui-icon ui-icon-transfer-e-w'
-            },
-            {
-                title: 'Update show in KODI',
-                path: `home/updateKODI?indexername=${indexerName}&seriesid=${showId}`,
-                requires: notifiers.kodi.enabled && notifiers.kodi.update.library,
-                icon: 'menu-icon-kodi'
-            },
-            {
-                title: 'Update show in Emby',
-                path: `home/updateEMBY?indexername=${indexerName}&seriesid=${showId}`,
-                requires: notifiers.emby.enabled,
-                icon: 'menu-icon-emby'
-            },
-            {
-                title: 'Preview Rename',
-                path: `home/testRename?indexername=${indexerName}&seriesid=${showId}`,
-                icon: 'ui-icon ui-icon-tag'
-            },
-            {
-                title: 'Download Subtitles',
-                path: `home/subtitleShow?indexername=${indexerName}&seriesid=${showId}`,
-                requires: config.subtitles.enabled && !isBeingSubtitled && show.config.subtitlesEnabled,
-                icon: 'menu-icon-backlog'
-            }
-        ]);
-    }
-    return menu;
-};
+import {
+    configSubMenu,
+    errorlogsSubMenu,
+    historySubMenu,
+    showSubMenu
+} from './sub-menus';
 
 const homeRoutes = [{
     path: '/home',
@@ -151,16 +81,6 @@ const homeRoutes = [{
     }
 }];
 
-const configSubMenu = [
-    { title: 'General', path: 'config/general/', icon: 'menu-icon-config' },
-    { title: 'Backup/Restore', path: 'config/backuprestore/', icon: 'menu-icon-backup' },
-    { title: 'Search Settings', path: 'config/search/', icon: 'menu-icon-manage-searches' },
-    { title: 'Search Providers', path: 'config/providers/', icon: 'menu-icon-provider' },
-    { title: 'Subtitles Settings', path: 'config/subtitles/', icon: 'menu-icon-backlog' },
-    { title: 'Post Processing', path: 'config/postProcessing/', icon: 'menu-icon-postprocess' },
-    { title: 'Notifications', path: 'config/notifications/', icon: 'menu-icon-notification' },
-    { title: 'Anime', path: 'config/anime/', icon: 'menu-icon-anime' }
-];
 const configRoutes = [{
     path: '/config',
     name: 'config',
@@ -329,10 +249,6 @@ const scheduleRoute = {
     }
 };
 
-const historySubMenu = [
-    { title: 'Clear History', path: 'history/clearHistory', icon: 'ui-icon ui-icon-trash', confirm: 'clearhistory' },
-    { title: 'Trim History', path: 'history/trimHistory', icon: 'menu-icon-cut', confirm: 'trimhistory' }
-];
 const historyRoute = {
     path: '/history',
     name: 'history',
@@ -409,39 +325,6 @@ const manageRoutes = [{
     }
 }];
 
-const errorlogsSubMenu = function() {
-    const { $route, $store } = this;
-    const level = $route.params.level || $route.query.level;
-    const { config } = $store.state;
-    const { loggingLevels, numErrors, numWarnings } = config.logs;
-    if (Object.keys(loggingLevels).length === 0) {
-        return [];
-    }
-
-    const isLevelError = (level === undefined || Number(level) === loggingLevels.error);
-
-    return [
-        {
-            title: 'Clear Errors',
-            path: 'errorlogs/clearerrors/',
-            requires: numErrors >= 1 && isLevelError,
-            icon: 'ui-icon ui-icon-trash'
-        },
-        {
-            title: 'Clear Warnings',
-            path: `errorlogs/clearerrors/?level=${loggingLevels.warning}`,
-            requires: numWarnings >= 1 && Number(level) === loggingLevels.warning,
-            icon: 'ui-icon ui-icon-trash'
-        },
-        {
-            title: 'Submit Errors',
-            path: 'errorlogs/submit_errors/',
-            requires: numErrors >= 1 && isLevelError,
-            confirm: 'submiterrors',
-            icon: 'ui-icon ui-icon-arrowreturnthick-1-n'
-        }
-    ];
-};
 const errorLogsRoutes = [{
     path: '/errorlogs',
     name: 'errorlogs',
